:recycle: refactor oracles related features
