+++ src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 
 import { GamesComponent } from '../features/games/components/games/games.component';
-import { OraclesComponent } from '../components/oracles/oracles.component';
+import { OraclesComponent } from '../features/oracles/components/oracles/oracles.component';
 import { BggSearchComponent } from '../components/bgg-search/bgg-search.component';
 
 export const routes: Routes = [
