+++ src/features/oracles/components/oracles/oracles.component.ts
@@ -11,14 +11,14 @@ import {
 import { CommonModule, NgOptimizedImage } from '@angular/common';
 import { MatButtonModule } from '@angular/material/button';
 
-import { CommonFunctionsService } from '../../core/functions/common/common-functions.service';
-import { FilterFunctionsService } from '../../core/functions/filter/filter-functions.service';
-import { HttpService } from '../../core/services/http/http.service';
-import { ScrollToTopBtnComponent } from '../scroll-to-top-btn/scroll-to-top-btn.component';
-import { OracleCard } from '../models';
-import { LoaderComponent } from '../loader/loader.component';
-import { LoaderService } from '../../core/services/loader/loader.service';
-import { ExportService } from '../../core/services/export/export.service';
+import { CommonFunctionsService } from '../../../../core/functions/common/common-functions.service';
+import { FilterFunctionsService } from '../../../../core/functions/filter/filter-functions.service';
+import { HttpService } from '../../../../core/services/http/http.service';
+import { ScrollToTopBtnComponent } from '../../../../components/scroll-to-top-btn/scroll-to-top-btn.component';
+import { OracleCard } from '../../../../components/models';
+import { LoaderComponent } from '../../../../components/loader/loader.component';
+import { LoaderService } from '../../../../core/services/loader/loader.service';
+import { ExportService } from '../../../../core/services/export/export.service';
 
 @Component({
   selector: 'app-oracles',
@@ -30,7 +30,7 @@ import { ExportService } from '../../core/services/export/export.service';
     ScrollToTopBtnComponent,
   ],
   templateUrl: './oracles.component.html',
-  styleUrl: '../common-styles.scss',
+  styleUrl: '../../../../components/common-styles.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class OraclesComponent implements OnInit, AfterViewInit {
